+++ resolvers/patient.js
@@ -17,10 +17,7 @@ const patientResolvers = {
       subscribe: withFilter(
         () => pubsub.asyncIterator([PATIENT_ADDED]),
         (payload, variables) => {
-          return (
-            payload.eventId ===
-            parseInt(variables.eventId)
-          );
+          return payload.eventId === parseInt(variables.eventId);
         }
       ),
     },
@@ -28,10 +25,7 @@ const patientResolvers = {
       subscribe: withFilter(
         () => pubsub.asyncIterator([PATIENT_UPDATED]),
         (payload, variables) => {
-          return (
-            payload.eventId ===
-            parseInt(variables.eventId)
-          );
+          return payload.eventId === parseInt(variables.eventId);
         }
       ),
     },
@@ -39,10 +33,7 @@ const patientResolvers = {
       subscribe: withFilter(
         () => pubsub.asyncIterator([PATIENT_RESTORED]),
         (payload, variables) => {
-          return (
-            payload.eventId ===
-            parseInt(variables.eventId)
-          );
+          return payload.eventId === parseInt(variables.eventId);
         }
       ),
     },
@@ -50,10 +41,7 @@ const patientResolvers = {
       subscribe: withFilter(
         () => pubsub.asyncIterator([PATIENT_DELETED]),
         (payload, variables) => {
-          return (
-            payload.eventId ===
-            parseInt(variables.eventId)
-          );
+          return payload.eventId === parseInt(variables.eventId);
         }
       ),
     },
@@ -118,8 +106,13 @@ const patientResolvers = {
         hospitalId: args.hospitalId,
         ambulanceId: args.ambulanceId,
       });
-      const collectionPoint = await db.collectionPoint.findByPk(newPatient.collectionPointId);
-      pubsub.publish(PATIENT_ADDED, { patientAdded: newPatient, eventId: collectionPoint.eventId });
+      const collectionPoint = await db.collectionPoint.findByPk(
+        newPatient.collectionPointId
+      );
+      pubsub.publish(PATIENT_ADDED, {
+        patientAdded: newPatient,
+        eventId: collectionPoint.eventId,
+      });
       return newPatient;
     },
     updatePatient: async (parent, args) => {
@@ -161,8 +154,13 @@ const patientResolvers = {
         }
       );
       const updatedPatient = await db.patient.findByPk(args.id);
-      const collectionPoint = await db.collectionPoint.findByPk(updatedPatient.collectionPointId);
-      pubsub.publish(PATIENT_UPDATED, { patientUpdated: updatedPatient, eventId: collectionPoint.eventId });
+      const collectionPoint = await db.collectionPoint.findByPk(
+        updatedPatient.collectionPointId
+      );
+      pubsub.publish(PATIENT_UPDATED, {
+        patientUpdated: updatedPatient,
+        eventId: collectionPoint.eventId,
+      });
       return updatedPatient;
     },
     restorePatient: async (parent, args) => {
@@ -175,8 +173,13 @@ const patientResolvers = {
         where: { id: args.id },
       });
       const restoredPatient = await db.patient.findByPk(args.id);
-      const collectionPoint = await db.collectionPoint.findByPk(restoredPatient.collectionPointId);
-      pubsub.publish(PATIENT_RESTORED, { patientRestored: restoredPatient, eventId: collectionPoint.eventId });
+      const collectionPoint = await db.collectionPoint.findByPk(
+        restoredPatient.collectionPointId
+      );
+      pubsub.publish(PATIENT_RESTORED, {
+        patientRestored: restoredPatient,
+        eventId: collectionPoint.eventId,
+      });
       return restoredPatient;
     },
     // This is a user delete of a patient, where the status is updated. A system delete happens if a CCP with associated patients is deleted
@@ -197,8 +200,13 @@ const patientResolvers = {
         .then(async (isDeleted) => {
           if (isDeleted[0] === 1) {
             const deletedPatient = await db.patient.findByPk(args.id);
-            const collectionPoint = await db.collectionPoint.findByPk(deletedPatient.collectionPointId);
-            pubsub.publish(PATIENT_DELETED, { patientDeleted: deletedPatient, eventId: collectionPoint.eventId });
+            const collectionPoint = await db.collectionPoint.findByPk(
+              deletedPatient.collectionPointId
+            );
+            pubsub.publish(PATIENT_DELETED, {
+              patientDeleted: deletedPatient,
+              eventId: collectionPoint.eventId,
+            });
             return deletedPatient;
           }
           throw new Error('Deletion failed for patient ID: ' + args.id);
